Fix import, allow the user to specify min and max values
