+++ python/accessible_graphs_package/accessible_graphs.egg-info/SOURCES.txt
@@ -5,4 +5,3 @@ accessible_graphs.egg-info/SOURCES.txt
 accessible_graphs.egg-info/dependency_links.txt
 accessible_graphs.egg-info/top_level.txt
 accessible_graphs_pkg/__init__.py
-accessible_graphs_pkg/accessibleGraphs.py
